Refactor: move part of `TokensTab` into `Form` component
